Flytt ut logikk som lager en boolsk-tidslinje med endringer i kompetanse pr person
